% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame.R
\name{frame}
\alias{frame}
\alias{hirs_frame_create}
\alias{hirs_frame_delete}
\alias{hirs_frame_updatetq}
\title{Frame operations}
\usage{
hirs_frame_create(index, frame, time_quantum = NULL, inverse_enabled = NULL,
  cache_type = "lru", cache_size = 50000L, range_enabled = NULL,
  fields = NULL, raw = FALSE, ...)

hirs_frame_delete(index, frame, raw = FALSE, ...)

hirs_frame_updatetq(index, frame, time_quantum, raw = FALSE, ...)
}
\arguments{
\item{index}{(character) A character vector of index names}

\item{frame}{(character) A frame name}

\item{time_quantum}{(character) Time Quantum for this frame}

\item{inverse_enabled}{(logical) Enables the inverted view for
this frame if \code{TRUE}}

\item{cache_type}{(character) ranked or LRU caching on this frame.
Default: "lru"}

\item{cache_size}{(integer) Number of rows to keep in the cache.
Default: 50,000}

\item{range_enabled}{(logical) Enables range-encoded fields in
this frame}

\item{fields}{(character) character vector of range-encoded fields}

\item{raw}{If \code{TRUE} (default), data is parsed to list. If \code{FALSE},
then raw JSON.}

\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Frame operations
}
\examples{
\dontrun{
# first, create an index
hirs_index_create("hello")

# then, create a frame
hirs_frame_create("hello", "world")
hirs_frame_delete("hello", "world")

# update frame time quantum
hirs_frame_create("hello", "world")
hirs_frame_updatetq("hello", "world", time_quantum = "MD")
}
}
