% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields.R
\name{hirs_field_create}
\alias{hirs_field_create}
\title{Fields}
\usage{
hirs_field_create(index, frame, field, type = "int", min, max, raw = FALSE,
  ...)
}
\arguments{
\item{index}{(character) A character vector of index names}

\item{frame}{(character) A frame name. required}

\item{field}{(character) A field name. required}

\item{type}{(character) A type, only "int". required}

\item{min}{(integer) minimum value. required}

\item{max}{(integer) maximum value. required}

\item{raw}{If \code{TRUE} (default), data is parsed to list. If \code{FALSE},
then raw JSON.}

\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Fields
}
\examples{
\dontrun{
# first, create an index
hirs_index_create("hello")

# then, create a frame
hirs_frame_create("hello", "world")
hirs_frame_delete("hello", "world")

# then, create a field
hirs_field_create(index="hello", frame="world", 
  field = "prs", type="int", min=0, max=100000L)
}
}
