% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hosts.R
\name{hirs_hosts}
\alias{hirs_hosts}
\title{List hosts}
\usage{
hirs_hosts(raw = FALSE, ...)
}
\arguments{
\item{raw}{If \code{TRUE} (default), data is parsed to list. If \code{FALSE},
then raw JSON.}

\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
List hosts
}
\examples{
\dontrun{
hirs_hosts()
}
}
