% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{hirs_schema}
\alias{hirs_schema}
\title{Get schema}
\usage{
hirs_schema(raw = FALSE, ...)
}
\arguments{
\item{raw}{If \code{TRUE} (default), data is parsed to list. If \code{FALSE},
then raw JSON.}

\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Get schema
}
\examples{
\dontrun{
hirs_schema()
}
}
