% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{hirs_search}
\alias{hirs_search}
\title{Search}
\usage{
hirs_search(index, body, column_attrs = NULL, slices = NULL, raw = FALSE,
  ...)
}
\arguments{
\item{index}{(character) A character vector of index names}

\item{body}{a named list passed to HTTP body}

\item{column_attrs}{column attributes}

\item{slices}{slices}

\item{raw}{If \code{TRUE} (default), data is parsed to list. If \code{FALSE},
then raw JSON.}

\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Search
}
\examples{
\dontrun{
hirs_search("repository", 'Bitmap(frame="language", rowID=5)')
hirs_search("repository", 'Bitmap(frame="language", rowID=5)', 
   columnAttrs = TRUE, slices = "0,1", verbose = TRUE)
hirs_search("repository", 'Count(Bitmap(frame="stargazer", rowID=10))')

# error
hirs_search("repository", 
  'Union(Bitmap(frame="stargazer", stargazer_id=1), Bitmap(frame="stargazer", stargazer_id=2))')
}
}
\references{
\url{https://www.pilosa.com/docs/latest/api-reference/}
}
