% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{hirs_version}
\alias{hirs_version}
\title{Get Pilosa version}
\usage{
hirs_version(raw = FALSE, ...)
}
\arguments{
\item{raw}{If \code{TRUE} (default), data is parsed to list. If \code{FALSE},
then raw JSON.}

\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Get Pilosa version
}
\examples{
\dontrun{
hirs_version()
}
}
