% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hirsutosa-package.r
\docType{package}
\name{hirsutosa}
\alias{hirsutosa}
\alias{hirsutosa-package}
\alias{hirsutosa-package}
\title{hirsutosa}
\description{
hirsutosa
}
\author{
Scott Chamberlain \email{myrmecocystus+r@gmail.com}
}
