% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{hirs_schema}
\alias{hirs_schema}
\title{Get schema}
\usage{
hirs_schema(...)
}
\arguments{
\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Get schema
}
\examples{
\dontrun{
hirs_schema()
}
}
