% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{hirs_index_get}
\alias{hirs_index_create}
\alias{hirs_index_delete}
\title{Index operations}
\usage{
hirs_index_get(index = NULL, raw = FALSE, ...)

hirs_index_create(index, raw = FALSE, ...)

hirs_index_delete(index, raw = FALSE, ...)
}
\arguments{
\item{index}{(character) A character vector of index names}

\item{raw}{If \code{TRUE} (default), data is parsed to list. If \code{FALSE},
then raw JSON.}

\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Index operations
}
\examples{
\dontrun{
hirs_index_get()
hirs_index_create("foobar")
hirs_index_delete("foobar")
}
}
\references{
<>
}
