% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index.R
\name{index}
\alias{index}
\alias{index_get}
\alias{index_create}
\alias{index_delete}
\title{Index API operations}
\usage{
index_get(index = NULL, raw = FALSE, ...)

index_create(index, raw = FALSE, ...)

index_delete(index, raw = FALSE, ...)
}
\arguments{
\item{index}{(character) A character vector of index names}

\item{raw}{If \code{TRUE} (default), data is parsed to list. If \code{FALSE},
then raw JSON.}

\item{...}{Curl args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\description{
Index API operations
}
\examples{
\dontrun{
index_get()
index_create("foobar")
index_delete("foobar")
}
}
\references{
<>
}
\author{
Scott Chamberlain \href{mailto:myrmecocystus@gmail.com}{myrmecocystus@gmail.com}
}
